# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Xpart App OpenApi API Documentation",
        "description": "Xpart App Using L5 Swagger OpenApi description",
        "contact": {
            "email": "schneidershades@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/admin/markup-pricing": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get all markup-pricing",
                "description": "Get all markup-pricing",
                "operationId": "allMarkupPricing",
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Post markup-pricing",
                "description": "Post markup-pricing",
                "operationId": "postUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MarkupPricingCreateFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/admin/users": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get all users",
                "description": "Get all users",
                "operationId": "allUsers",
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Post users",
                "description": "Post users",
                "operationId": "postUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserCreateFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MarkupPricingCreateFormRequest": {
                "title": "Markup Pricing Create Form Request Fields",
                "description": "Markup Pricing Create request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "min_value": {
                        "title": "Max Value",
                        "description": "Maximum Value of the Price",
                        "type": "string",
                        "example": "10000.00"
                    },
                    "max_value": {
                        "title": "Min Value",
                        "description": "Minimum Value of the Price",
                        "type": "string",
                        "example": "100000.00"
                    },
                    "percentage": {
                        "title": "Percentage Value",
                        "description": "Percentage Value of the Price",
                        "type": "string",
                        "example": "100"
                    },
                    "type": {
                        "title": "Markup Type",
                        "description": "Type of the Price",
                        "type": "boolean",
                        "example": "quotes"
                    }
                },
                "type": "object"
            },
...
            "UserCreateFormRequest": {
                "title": "User Create Form Request Fields",
                "description": "User Create request body data",
                "required": [
                    "name"
                ],
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}